+++ src/main/java/de/cowtipper/cowlection/listener/PlayerListener.java
@@ -241,7 +241,7 @@ public class PlayerListener {
     public static boolean registerSkyBlockListeners() {
         if (dungeonsListener == null && skyBlockListener == null) {
             MinecraftForge.EVENT_BUS.register(dungeonsListener = new DungeonsListener(Cowlection.getInstance()));
-            MinecraftForge.EVENT_BUS.register(skyBlockListener = new SkyBlockListener());
+            MinecraftForge.EVENT_BUS.register(skyBlockListener = new SkyBlockListener(Cowlection.getInstance()));
             return true;
         }
         return false;