+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - SkyBlock Bazaar graphs improvements:
   - make graphs easier to read by connecting graphs' nodes
   - Fix graphs when using the mc unicode font
+- Added quick lookup for item prices and wiki:
+  - wiki: hypixel-skyblock.fandom.com (default key: `I` = info)
+  - item prices: stonks.gg (default key: `P` = price)
 
 ### Changed
 - Item age: show timestamp in the local timezone instead of "SkyBlock"-timezone (Eastern Time; also fixed the incorrect 12h ↔ 24h clock conversion)